Corrected 61850-90-5 timestamp encoding.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS07]$/gsf/Main;C83611

+++ Source/Libraries/GSF.PhasorProtocols/Iec61850_90_5/DataFrame.cs
@@ -449,7 +449,7 @@ namespace GSF.PhasorProtocols.Iec61850_90_5
                 m_configurationRevision.EncodeTagValue(SampledValueTag.ConfRev, asduImage, ref index);
 
                 // Encode timestamp
-                ulong timestamp = Word.MakeQword(header.SecondOfCentury, (uint)(header.FractionOfSecond << 8 | (int)header.TimeQualityFlags));
+                ulong timestamp = Word.MakeQword(header.SecondOfCentury, (uint)(((uint)header.FractionOfSecond << 8) | (uint)header.TimeQualityFlags));
                 timestamp.EncodeTagValue(SampledValueTag.RefrTm, asduImage, ref index);
 
                 // Defaulting sample synchronization state to true - not sure what value this has
